JD-172 save

+++ lib/app/plan/plans/diamond.php
@@ -18,7 +18,7 @@ class diamond implements plan
     }
 
 
-    public function priceRial(): int
+    public function priceRial(): float
     {
         return 900000; // IRT
     }
